Fix: counter: cast UINT*_MAX to 64-bit signed type before negative
(long long)-UINT32_MAX leads to value 1 which is not what we expect.
This is due to implicit type promotion from unsigned to signed 32-bit
integer.

Apply this to 8-bit and 16-bit types as well even though they are
not affected by this issue to keep things regular.

+++ libcounter/counter-api.h
@@ -104,7 +104,7 @@ static inline int __lttng_counter_add(const struct lib_counter_config *config,
 		}
 		if (v > 0 && (v >= UINT8_MAX || n < old))
 			overflow = true;
-		else if (v < 0 && (v <= -UINT8_MAX || n > old))
+		else if (v < 0 && (v <= -(int64_t) UINT8_MAX || n > old))
 			underflow = true;
 		break;
 	}
@@ -143,7 +143,7 @@ static inline int __lttng_counter_add(const struct lib_counter_config *config,
 		}
 		if (v > 0 && (v >= UINT16_MAX || n < old))
 			overflow = true;
-		else if (v < 0 && (v <= -UINT16_MAX || n > old))
+		else if (v < 0 && (v <= -(int64_t) UINT16_MAX || n > old))
 			underflow = true;
 		break;
 	}
@@ -182,7 +182,7 @@ static inline int __lttng_counter_add(const struct lib_counter_config *config,
 		}
 		if (v > 0 && (v >= UINT32_MAX || n < old))
 			overflow = true;
-		else if (v < 0 && (v <= -UINT32_MAX || n > old))
+		else if (v < 0 && (v <= -(int64_t) UINT32_MAX || n > old))
 			underflow = true;
 		break;
 	}
